Prepare for heroku
