+++ .gitignore
@@ -1,3 +1,4 @@
+target
 *.class
 
 # Mobile Tools for Java (J2ME)
