+++ src/main/java/ru/youcon/ottepel/model/TelegramUser.java
@@ -1,0 +1,54 @@
+package ru.youcon.ottepel.model;
+
+import com.asana.models.User;
+
+/**
+ * Created by savetisyan on 17/03/17
+ */
+public class TelegramUser {
+    private long id;
+    private String token;
+    private User user;
+
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public String getToken() {
+        return token;
+    }
+
+    public void setToken(String token) {
+        this.token = token;
+    }
+
+    public User getUser() {
+        return user;
+    }
+
+    public void setUser(User user) {
+        this.user = user;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+
+        TelegramUser that = (TelegramUser) o;
+
+        if (id != that.id) return false;
+        return user != null ? user.equals(that.user) : that.user == null;
+    }
+
+    @Override
+    public int hashCode() {
+        int result = (int) (id ^ (id >>> 32));
+        result = 31 * result + (user != null ? user.hashCode() : 0);
+        return result;
+    }
+}
