--- telegram-bot/src/main/java/ru/ottepel/Main.java
@@ -1,12 +1,0 @@
-package ru.ottepel;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.telegram.telegrambots.exceptions.TelegramApiRequestException;
-
-@SpringBootApplication(scanBasePackages = "ru.ottepel")
-public class Main {
-    public static void main(String[] args) throws TelegramApiRequestException {
-        SpringApplication.run(Main.class, args);
-    }
-}
