--- telegram-bot/src/main/resources/application.yml
@@ -1,5 +1,0 @@
-telegram:
-  botName: AsanaOttepelBot
-  botToken: 338480209:AAF9L-zDYHfJTso4tpQWMuXDZvdsLrABcH0
-  template: haha
-host: https://4400c0c4.ngrok.io
